Move eligibility items to their own flat package.
