+++ CorePayments/src/main/java/com/paypal/android/corepayments/api/eligibility/FundingEligibilityIntent.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.corepayments.graphql.fundingEligibility.models
+package com.paypal.android.corepayments.api.eligibility
 
 internal enum class FundingEligibilityIntent {
     SALE, CAPTURE, AUTHORIZE, ORDER, TOKENIZE, SUBSCRIPTION
