+++ CorePayments/src/main/java/com/paypal/android/corepayments/api/eligibility/SupportedCountryCurrencyType.kt
@@ -1,0 +1,5 @@
+package com.paypal.android.corepayments.api.eligibility
+
+internal enum class SupportedCountryCurrencyType {
+    USD
+}
