--- CorePayments/src/main/java/com/paypal/android/corepayments/graphql/fundingEligibility/models/SupportedCountryCurrencyType.kt
@@ -1,5 +1,0 @@
-package com.paypal.android.corepayments.graphql.fundingEligibility.models
-
-internal enum class SupportedCountryCurrencyType {
-    USD
-}
